Reorganize git CLI commands
