+++ src/cmd/git/clone.ts
@@ -1,0 +1,74 @@
+import { Command } from "@cliffy/command";
+import sdk, { user } from "~/sdk.ts";
+import { DEFAULT_BRANCH_NAME, DEFAULT_IGNORE_PATTERNS } from "~/consts.ts";
+import { parseProjectUri } from "~/cmd/parsing.ts";
+import VTClient from "~/vt/vt/VTClient.ts";
+import Kia from "kia";
+import { checkDirectory } from "~/utils.ts";
+import { basename } from "@std/path";
+import * as join from "@std/path/join";
+
+export const cloneCmd = new Command()
+  .name("clone")
+  .description("Clone a val town project")
+  .arguments("<projectUri:string> [cloneDir:string] [branchName:string]")
+  .example(
+    "Clone with username/projectName",
+    `vt clone username/projectName`,
+  )
+  .example(
+    "Clone with link",
+    `vt clone https://www.val.town/x/username/projectName`,
+  )
+  .action(
+    async (
+      _: unknown,
+      projectUri: string,
+      rootPath?: string,
+      branchName?: string,
+    ) => {
+      const spinner = new Kia("Cloning project...");
+      let targetDir = rootPath || Deno.cwd();
+
+      try {
+        const { ownerName, projectName } = parseProjectUri(
+          projectUri,
+          user.username!,
+        );
+
+        branchName = branchName || DEFAULT_BRANCH_NAME;
+
+        // By default, if the target directory is the current working directory,
+        // then use the project name as the target directory
+        if (rootPath === undefined) {
+          targetDir = join.join(targetDir, projectName);
+        }
+
+        const vt = await VTClient.init(
+          targetDir,
+          ownerName,
+          projectName,
+          undefined,
+          branchName,
+        );
+
+        // Make sure that the directory is safe to clone into (exists, or gets
+        // created and then exists, and wasn't nonempty)
+        await checkDirectory(targetDir, {
+          ignoreGlobs: DEFAULT_IGNORE_PATTERNS,
+        });
+
+        spinner.start();
+        await vt.clone(targetDir);
+        spinner.succeed(
+          `Project ${ownerName}/${projectName} cloned to ${
+            "./" + basename(targetDir)
+          }`,
+        );
+      } catch (error) {
+        if (error instanceof Error) {
+          spinner.fail(error.message);
+        }
+      }
+    },
+  );
