+++ src/cmd/git/index.ts
@@ -1,0 +1,5 @@
+export { cloneCmd } from "./clone.ts";
+export { pullCmd } from "./pull.ts"; 
+export { statusCmd } from "./status.ts";
+export { branchCmd } from "./branch.ts";
+export { checkoutCmd } from "./checkout.ts";
